pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-test-harness"
version = "0.3.0"
description = "Security and conformance testing for MCP servers"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "jsonschema>=4.20.0",
    "mcp>=1.20,<2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
mcp-test-harness = "mcp_test_harness.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_test_harness*"]
